Move weaken/strengthen/Extends into Tile.Functor.Compose.
